Moved order book related components to components/order-book
