+++ src/components/order-book/OrderBookTradesWidget/OrderBookTrades.styled.ts
